StudentHelper/requirements_.txt
aiohttp==3.8.4
beautifulsoup4==4.12.2
cffi==1.15.1
Django==3.2.10
djangorestframework==3.11.0
docx==0.2.4
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
lxml==4.9.2
multidict==6.0.4
openai==0.27.8
outcome==1.2.0
psycopg2==2.9.6
pycparser==2.21
pyOpenSSL==23.2.0
python-docx==0.8.11
pytz==2023.3
requests==2.31.0
selenium==4.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
